Добавил обработку ошибок при покупке и удалении ключевых зданий: ратуша и шахта

+++ server/api/application/village/Village.php
@@ -71,10 +71,16 @@ class Village {
         if (!$village) {
             return ['error' => 310];
         }
+
         $building = $this->db->getBuildingType($typeId);
         if (!$building) {
             return ["error" => 301];
         }
+
+        if ($building->id === '1' || $building->id === '2') {
+            return ['error' => 307];
+        }
+        
         if ($user->money < $building->price) {
             return ['error' => 305];
         }
@@ -138,6 +144,10 @@ class Village {
         $building = $this->db->getBuilding($buildingId, $village->id);
         if (!$building) {
             return ['error' => 300];
+        }
+
+        if ($building->typeId === '1' || $building->typeId === '2') {
+            return ['error' => 307];
         }
 
         $result = $this->db->deleteBuilding($buildingId, $village->id);
